+++ samples/web_api_demo.py
@@ -1,4 +1,6 @@
-from ami.threads import MainThread
+from ami.threads.thread_control import ThreadController
+from ami.threads.web_api_handler import WebApiHandler
 
-main_thread = MainThread()
-main_thread.start()
+controller = ThreadController()
+handler = WebApiHandler(controller)
+handler.run()
